# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Qwen API Gateway Wrapper
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ToolDefinitionFunctionParameters = { [key: string]: unknown };

export type ToolDefinitionFunction = {
  name: string;
  description?: string;
  parameters?: ToolDefinitionFunctionParameters;
};

export interface ToolDefinition {
  type: string;
  function: ToolDefinitionFunction;
}

export type ChatMessageToolCallsItem = { [key: string]: unknown };

export interface ChatMessage {
  role: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  tool_calls?: ChatMessageToolCallsItem[] | null;
  /** @nullable */
  tool_call_id?: string | null;
}

/**
 * The request body to forward
 */
export type ProxyRequestInputPayload = { [key: string]: unknown };

/**
 * Additional headers to pass through
 */
export type ProxyRequestInputExtraHeaders = { [key: string]: unknown };

export interface ProxyRequestInput {
  /** Bearer token for Qwen API authentication */
  token: string;
  /** The endpoint path to proxy to (e.g. chat/completions) */
  endpoint: string;
  /** HTTP method */
  method: string;
  /** The request body to forward */
  payload?: ProxyRequestInputPayload;
  /** Additional headers to pass through */
  extraHeaders?: ProxyRequestInputExtraHeaders;
}

export type ProxyResponseRequestPayload = { [key: string]: unknown };

export type ProxyResponseResponseBody = { [key: string]: unknown };

export type ProxyResponseResponseHeaders = { [key: string]: unknown };

export interface ProxyResponse {
  id: string;
  success: boolean;
  statusCode: number;
  requestedAt: string;
  responseTime: number;
  endpoint: string;
  method: string;
  requestPayload?: ProxyResponseRequestPayload;
  responseBody?: ProxyResponseResponseBody;
  responseHeaders?: ProxyResponseResponseHeaders;
  /** @nullable */
  error?: string | null;
}

export type HistoryEntryRequestPayload = { [key: string]: unknown };

export type HistoryEntryResponseBody = { [key: string]: unknown };

export interface HistoryEntry {
  id: string;
  success: boolean;
  statusCode: number;
  requestedAt: string;
  responseTime: number;
  endpoint: string;
  method: string;
  requestPayload?: HistoryEntryRequestPayload;
  responseBody?: HistoryEntryResponseBody;
  /** @nullable */
  error?: string | null;
}

export interface GatewayStats {
  totalRequests: number;
  successCount: number;
  failureCount: number;
  avgResponseTime: number;
  /** @nullable */
  lastRequestAt: string | null;
}

export interface ClearResult {
  cleared: number;
}

export type GetHistoryParams = {
/**
 * Max number of records to return
 */
limit?: number;
};