# lib/api-zod/src/generated/types/proxyRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Qwen API Gateway Wrapper
 * OpenAPI spec version: 0.1.0
 */
import type { ProxyRequestInputExtraHeaders } from './proxyRequestInputExtraHeaders';
import type { ProxyRequestInputPayload } from './proxyRequestInputPayload';

export interface ProxyRequestInput {
  /** Bearer token for Qwen API authentication */
  token: string;
  /** The endpoint path to proxy to (e.g. chat/completions) */
  endpoint: string;
  /** HTTP method */
  method: string;
  /** The request body to forward */
  payload?: ProxyRequestInputPayload;
  /** Additional headers to pass through */
  extraHeaders?: ProxyRequestInputExtraHeaders;
}